+++ NEWS.md
@@ -25,6 +25,16 @@ dispRity 0.2.0 (2016-04-01)
 <!--   * `dtt.dispRity` now works with any tree (including non-ultrametric ones and fossils only ones). -->
 <!--   * `dtt.dispRity` now works with time-slicing. -->
 
+
+dispRity v1.2.3 (2019-03-12)
+=========================
+
+### MINOR IMPROVEMENTS
+
+ * `centroids` and `ancestral.dist` functions can now take the `method` option for `"euclidean"` or `"manhattan"` distances.
+ 
+<!-- * All functions methods selection have now been sped up using `switch`. -->
+
 dispRity v1.2.3 (2019-03-12)
 =========================